test(logic): trim 10 oversized tests via splits and shared helpers
Refs waigore/colonizethis#2216

- Add combatGame and unit shorthand factories on TestFixtures; use helpers in player_view tests
- Extract turn news digest builders and civilian-build topology helpers into *_test_support
- Centralize purchase-land fixture graph in order_engine_purchase_land_test_support; split validators across two test files
- Move tail scenarios from army, naval, sim AI, game setup landmass, build improvement, suggestion, civilian build validation, and spy intel PlayerView coverage into sibling *_test dart files under 400 lines
- Remove unused game_setup landmass P–P adjacency helper previously dead code
