+++ packages/colonizethis_logic/test/order_engine_validate_work_build_improvement_test.dart
@@ -381,79 +381,6 @@ void main() {
           contains('foreign or uncontrolled province'),
         );
       });
-
-      test(
-        'accepts build_improvement on purchased tile in foreign province',
-        () {
-          final foreignProvinceId = '$ow|P2';
-          final foreignTileKey = '$foreignProvinceId|0|0';
-          final game = Game(
-            id: 'g1',
-            worldState: WorldState(
-              turnState: const TurnState(
-                phase: TurnPhase.orders,
-                turnNumber: 0,
-              ),
-              oldWorld: RegionData(
-                provinces: [
-                  Province(id: provinceId, regionId: ow, ownerId: 'p1'),
-                  Province(id: foreignProvinceId, regionId: ow, ownerId: 'p2'),
-                ],
-                units: [
-                  Unit(
-                    id: 'builder1',
-                    type: kUnitTypeBuilder,
-                    ownerId: 'p1',
-                    locationProvinceId: provinceId,
-                    tileKey: tileKey,
-                  ),
-                ],
-              ),
-              newWorld: const RegionData(),
-              resourceByTileKey: {tileKey: 'grain', foreignTileKey: 'grain'},
-              tileState: const TileMapState(),
-              tileKeysByRegionAndProvince: {
-                ow: {
-                  provinceId: [tileKey],
-                  foreignProvinceId: [foreignTileKey],
-                },
-              },
-              playerVisibilityByTile: {
-                'p1': {tileKey: 'fullyVisible', foreignTileKey: 'fullyVisible'},
-              },
-              purchasedTilesByTileKey: {foreignTileKey: 'p1'},
-            ),
-            players: [
-              Player(
-                id: 'p1',
-                displayName: 'P1',
-                isHuman: true,
-                capitalProvinceId: provinceId,
-                stockpile: Stockpile()
-                    .applyDelta(CommodityCatalog.lumber.id, 2)
-                    .applyDelta(CommodityCatalog.castIron.id, 2),
-                techUnlocked: const {kTechIdCircularSaw: true},
-              ),
-              const Player(id: 'p2', displayName: 'P2', isHuman: false),
-            ],
-          );
-          final engine = OrderEngine();
-          engine.addWorkOrder(
-            'p1',
-            WorkOrder(
-              unitId: 'builder1',
-              target: kWorkTargetBuildImprovement,
-              targetTileKey: foreignTileKey,
-            ),
-          );
-          final results = engine.validatePlayerOrdersWithContext(
-            game,
-            topology,
-            'p1',
-          );
-          expect(results.single.status, OrderValidationStatus.accepted);
-        },
-      );
     });
   });
 }
