+++ packages/colonizethis_logic/test/order_suggestion_core_part2_test.dart
@@ -378,48 +378,5 @@ void main() {
         );
       },
     );
-
-    test('suggestNavalMissionOrders returns list', () {
-      const playerId = 'gp1';
-      const ow = 'oldWorld';
-      final player = const Player(
-        id: playerId,
-        displayName: 'GP',
-        isHuman: false,
-      );
-      final world = WorldState(
-        turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-        oldWorld: const RegionData(),
-        newWorld: const RegionData(),
-        fleets: [
-          Fleet(
-            id: 'fleet_gp1',
-            ownerId: playerId,
-            seaZoneId: 'sea1',
-            regionId: ow,
-            shipTypeIds: ['fluyte'],
-          ),
-        ],
-      );
-      final game = Game(id: 'g1', worldState: world, players: [player]);
-      final topology = MapTopology(
-        nodes: const [
-          TopologyNode(
-            id: 'sea1',
-            regionId: ow,
-            type: TopologyNodeType.seaZone,
-          ),
-        ],
-        edges: const [],
-      );
-      final view = buildPlayerView(game, topology, playerId);
-      final suggestions = suggestNavalMissionOrders(
-        view,
-        game,
-        topology,
-        const Orders(),
-      );
-      expect(suggestions, isA<List<NavalMissionOrder>>());
-    });
   });
 }
