+++ packages/colonizethis_logic/test/player_view_test.dart
@@ -3,6 +3,8 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
+import 'test_fixtures.dart';
+
 void main() {
   group('PlayerView', () {
     test('buildPlayerView collects own units, provinces and diplomacy', () {
@@ -26,13 +28,13 @@ void main() {
         ownerId: 'gp2',
       );
 
-      final u1 = Unit(
+      final u1 = TestFixtures.testMilitaryUnit(
         id: 'u1',
         type: 'inf',
         ownerId: 'gp1',
         locationProvinceId: '$ow|p1',
       );
-      final u2 = Unit(
+      final u2 = TestFixtures.testMilitaryUnit(
         id: 'u2',
         type: 'inf',
         ownerId: 'gp2',
@@ -378,51 +380,5 @@ void main() {
       },
     );
 
-    test(
-      'province panel intel shows for foreign province with active spy timer',
-      () {
-        const ow = 'oldWorld';
-        const provinceId = '$ow|p2';
-        const tileKey = '$provinceId|0|0';
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-            oldWorld: RegionData(
-              provinces: [
-                Province(id: provinceId, regionId: ow, ownerId: 'gp2'),
-              ],
-            ),
-            newWorld: const RegionData(),
-            tileKeysByRegionAndProvince: {
-              ow: {
-                provinceId: [tileKey],
-              },
-            },
-            playerVisibilityByTile: const {
-              'gp1': {tileKey: 'fogged'},
-            },
-            spyRevealTurnsByPlayer: const {
-              'gp1': {provinceId: 2},
-            },
-          ),
-          players: const [
-            Player(id: 'gp1', displayName: 'GP1', isHuman: true),
-            Player(id: 'gp2', displayName: 'GP2', isHuman: false),
-          ],
-        );
-        const topology = MapTopology(nodes: [], edges: []);
-        final view = buildPlayerView(game, topology, 'gp1');
-        expect(
-          provincePanelShowsFullTileDerivedIntel(
-            game: game,
-            view: view,
-            humanPlayerId: 'gp1',
-            provinceId: provinceId,
-          ),
-          isTrue,
-        );
-      },
-    );
   });
 }
